Fix: promo controlador para crear promos y formcrearcliente , tipo dedato del dni de int a long

+++ Eterea_Parfums_Desktop/FormABM/Cliente/FormCrearCliente.cs
@@ -317,7 +317,7 @@ namespace Eterea_Parfums_Desktop
 
 
             Cliente cliente = new Cliente(0, txt_usuario.Text, txt_clave.Text, txt_nombre.Text, txt_apellido.Text,
-                int.Parse(txt_dni.Text), combo_con_iva.Text, DateTime.Parse(dateTime_nac.Text), txt_celular.Text, txt_email.Text,
+                long.Parse(txt_dni.Text), combo_con_iva.Text, DateTime.Parse(dateTime_nac.Text), txt_celular.Text, txt_email.Text,
                 pais, provincia, ciudad, int.Parse(txt_cp.Text), calle, int.Parse(txt_num_calle.Text),
                 txt_piso.Text, txt_depto.Text, richTextBox_comentario.Text,
                  activo, rol);
